refactor:refactor Logo and Navigation like in develop
